+++ RGB.NET.Core/Devices/IRGBDevice.cs
@@ -43,7 +43,7 @@ namespace RGB.NET.Core
         /// <param name="referenceRect">The <see cref="Rectangle"/> to check.</param>
         /// <param name="minOverlayPercentage">The minimal percentage overlay a <see cref="Led"/> must have with the <see cref="Rectangle" /> to be taken into the list.</param>
         /// <returns></returns>
-        IEnumerable<Led> this[Rectangle referenceRect, float minOverlayPercentage = 0.5f] { get; }
+        IEnumerable<Led> this[Rectangle referenceRect, double minOverlayPercentage = 0.5] { get; }
 
         #endregion
